# Conflicts:
#	src/main/java/com/kgc/easybuy/aop/Logg.java
#	src/main/java/com/kgc/easybuy/controller/CatController.java
#	src/main/java/com/kgc/easybuy/controller/CategoryController.java
#	src/main/java/com/kgc/easybuy/controller/OrderController.java
#	src/main/java/com/kgc/easybuy/controller/ProductController.java
#	src/main/java/com/kgc/easybuy/dao/CatMapper.java
#	src/main/java/com/kgc/easybuy/dao/OrderMapper.java
#	src/main/java/com/kgc/easybuy/dao/Order_detailMapper.java
#	src/main/java/com/kgc/easybuy/dao/ProductMapper.java
#	src/main/java/com/kgc/easybuy/dao/UserMapper.java
#	src/main/java/com/kgc/easybuy/pojo/Cat.java
#	src/main/java/com/kgc/easybuy/pojo/Product.java
#	src/main/java/com/kgc/easybuy/pojo/User.java
#	src/main/java/com/kgc/easybuy/service/CatService.java
#	src/main/java/com/kgc/easybuy/service/OrderService.java
#	src/main/java/com/kgc/easybuy/service/ProductService.java
#	src/main/java/com/kgc/easybuy/service/UserService.java
#	src/main/java/com/kgc/easybuy/service/impl/CatServiceImpl.java
#	src/main/java/com/kgc/easybuy/service/impl/OrderServiceImpl.java
#	src/main/java/com/kgc/easybuy/service/impl/ProductServiceImpl.java
#	src/main/java/com/kgc/easybuy/service/impl/UserServiceImpl.java
#	src/main/resources/mapper/CategoryMapper.xml
#	src/main/resources/mapper/OrderMapper.xml
#	src/main/resources/mapper/Order_detailMapper.xml
#	src/main/resources/mapper/ProductMapper.xml
#	src/main/resources/mapper/UserMapper.xml
#	src/main/resources/static/CategoryList.html
#	src/main/resources/static/index.html
#	src/main/resources/static/order/user_order.html
#	src/main/resources/static/product_update.html
#	target/classes/static/member/Member_Address.html
#	target/classes/static/member/Member_Address_Update.html

+++ src/main/java/com/kgc/easybuy/pojo/AlipayDBPojo.java
@@ -12,8 +12,9 @@ import java.io.Serializable;
 @AllArgsConstructor
 @NoArgsConstructor
 public class AlipayDBPojo implements Serializable {
+
     private String out_trade_no;//订单号
-    private int total_amount;//总金额
+    private double total_amount;//总金额
     private int status;//状态
